Release 3.1.0 RC1, 2012-08-22

Conflicts:
	alljoyn_android/SConscript
	alljoyn_android/alljoyn/AndroidManifest.xml
	alljoyn_android/alljoyn/jni/Android.mk
	alljoyn_android/alljoyn/src/org/alljoyn/bus/alljoyn/AllJoynApp.java
	alljoyn_android/alljoyn_android_ext/.classpath
	alljoyn_android/alljoyn_android_ext/src/org/alljoyn/jni/StartServiceReceiver.java
	alljoyn_android/alljoyn_android_p2pservice/.project
	alljoyn_android/bundle_test/BundledClient/AndroidManifest.xml
	alljoyn_android/bundle_test/BundledService/AndroidManifest.xml
	alljoyn_android/ice_client_login_service/AndroidManifest.xml
	alljoyn_android/ice_client_login_service/res/values/strings.xml
	alljoyn_android/ice_client_login_service/src/org/alljoyn/ice/ClientLoginService.java
	alljoyn_android/ice_client_login_service/src/org/alljoyn/ice/ClientLoginServiceInterface.java
	daemon/winrt/proximity/ProximityListener.h
	daemon/winrt/proximity/SConscript
	samples/simple/android/client/jni/Android.mk
	samples/simple/android/service/jni/Android.mk
	samples/windows/Client/Client.cpp
	samples/windows/Client/Client.vcproj
	samples/windows/PhotoChat/AllJoynBusLib/AllJoynBusAPI.h
	samples/windows/PhotoChat/AllJoynBusLib/AllJoynBusLib.cpp
	samples/windows/PhotoChat/AllJoynBusLib/AllJoynBusLib.vcproj
	samples/windows/PhotoChat/AllJoynBusLib/AllJoynConnection.cpp
	samples/windows/PhotoChat/AllJoynBusLib/AllJoynConnection.h
	samples/windows/PhotoChat/AllJoynBusLib/dllmain.cpp
	samples/windows/PhotoChat/AllJoynBusLib/stdafx.cpp
	samples/windows/PhotoChat/AllJoynBusLib/stdafx.h
	samples/windows/PhotoChat/AllJoynBusLib/targetver.h
	samples/windows/PhotoChat/AllJoynNET/AJBus.cs
	samples/windows/PhotoChat/AllJoynNET/AllJoynNET.csproj
	samples/windows/PhotoChat/AllJoynNET/AlljoynConnect.Designer.cs
	samples/windows/PhotoChat/AllJoynNET/AlljoynConnect.cs
	samples/windows/PhotoChat/AllJoynNET/AlljoynConnect.resx
	samples/windows/PhotoChat/AllJoynNET/AlljoynConnectForm.cs
	samples/windows/PhotoChat/AllJoynNET/AlljoynConnectForm.designer.cs
	samples/windows/PhotoChat/AllJoynNET/AlljoynConnectForm.resx
	samples/windows/PhotoChat/AllJoynNET/AlljoynSetup.Designer.cs
	samples/windows/PhotoChat/AllJoynNET/AlljoynSetup.cs
	samples/windows/PhotoChat/AllJoynNET/AlljoynSetup.resx
	samples/windows/PhotoChat/AllJoynNET/Properties/AssemblyInfo.cs
	samples/windows/PhotoChat/AllJoynNET/Properties/Resources.Designer.cs
	samples/windows/PhotoChat/AllJoynNET/Properties/Resources.resx
	samples/windows/PhotoChat/AllJoynNET/Properties/Settings.Designer.cs
	samples/windows/PhotoChat/AllJoynNET/Properties/Settings.settings
	samples/windows/PhotoChat/AllJoynNET/SimpleChatForm.Designer.cs
	samples/windows/PhotoChat/AllJoynNET/SimpleChatForm.cs
	samples/windows/PhotoChat/AllJoynNET/SimpleChatForm.resx
	samples/windows/PhotoChat/AllJoynNET/SimpleTranscript.Designer.cs
	samples/windows/PhotoChat/AllJoynNET/SimpleTranscript.resx
	samples/windows/PhotoChat/AllJoynNET/StartForm.Designer.cs
	samples/windows/PhotoChat/AllJoynNET/StartForm.resx
	samples/windows/PhotoChat/AllJoynNET/Transcript.Designer.cs
	samples/windows/PhotoChat/AllJoynNET/Transcript.cs
	samples/windows/PhotoChat/AllJoynNET/Transcript.resx
	samples/windows/PhotoChat/AllJoynNET/logo.gif
	samples/windows/PhotoChat/PhotoChat.csproj
	samples/windows/PhotoChat/PhotoChat.sln
	samples/windows/PhotoChat/PhotoChatForm.Designer.cs
	samples/windows/PhotoChat/PhotoChatForm.cs
	samples/windows/PhotoChat/PhotoChatForm.resx
	samples/windows/PhotoChat/Properties/AssemblyInfo.cs
	samples/windows/PhotoChat/Properties/Resources.Designer.cs
	samples/windows/PhotoChat/Properties/Resources.resx
	samples/windows/PhotoChat/Properties/Settings.Designer.cs
	samples/windows/PhotoChat/Properties/Settings.settings
	samples/windows/PhotoChat/RichTextBuffer.cs
	samples/windows/PhotoChat/logo.gif
	samples/windows/Service/Service.cpp
	samples/windows/Service/Service.vcproj
	samples/windows/SetUserEnvironmentVariable.vbs
	samples/windows/chat/AlljoynSetup.Designer.cs
	samples/windows/chat/AlljoynSetup.cs
	samples/windows/chat/AlljoynSetup.resx
	samples/windows/chat/Chat.sln
	samples/windows/chat/ChatDialog.Designer.cs
	samples/windows/chat/ChatDialog.cs
	samples/windows/chat/ChatDialog.resx
	samples/windows/chat/ChatLib32/ChatClasses.cpp
	samples/windows/chat/ChatLib32/ChatClasses.h
	samples/windows/chat/ChatLib32/ChatLib32.cpp
	samples/windows/chat/ChatLib32/ChatLib32.h
	samples/windows/chat/ChatLib32/ChatLib32.vcproj
	samples/windows/chat/Program.cs
	samples/windows/chat/Properties/AssemblyInfo.cs
	samples/windows/chat/Properties/Resources.Designer.cs
	samples/windows/chat/Properties/Resources.resx
	samples/windows/chat/Properties/Settings.Designer.cs
	samples/windows/chat/Properties/Settings.settings
	samples/windows/chat/chat.csproj
	samples/windows/chat/logo.gif
	src/winrt/binding/inc/AllJoynException.h
	src/winrt/binding/inc/SocketStreamEvent.h

+++ unit_test/ClientSetup.cc
@@ -183,7 +183,7 @@ QStatus ClientSetup::MethodCall(int noOfCalls, int type)
         const MsgArg*replyArgs[10];
         size_t numArgs;
         reply->GetArgs(numArgs, *replyArgs);
-        EXPECT_EQ(10, numArgs);
+        EXPECT_EQ(static_cast<size_t>(10), numArgs);
     }     // end if type == 3
     else if (type == 4) {
 
@@ -383,12 +383,12 @@ void ClientSetup::MySignalHandler2(
     const MsgArg*replyArgs[2];
     size_t numArgs;
     msg->GetArgs(numArgs, *replyArgs);
-    EXPECT_EQ(2, numArgs);
+    EXPECT_EQ(static_cast<size_t>(2), numArgs);
     if (replyArgs[0]->v_uint32 == 5) {
         EXPECT_STREQ("hello", msg->GetArg(1)->v_string.str);
-        EXPECT_EQ(5, msg->GetArg(1)->v_string.len);
+        EXPECT_EQ(static_cast<uint32_t>(5), msg->GetArg(1)->v_string.len);
     } else if (msg->GetArg(1)->v_uint32 == 4096) {
-        EXPECT_EQ(4096, msg->GetArg(1)->v_string.len);
+        EXPECT_EQ(static_cast<uint32_t>(4096), msg->GetArg(1)->v_string.len);
         qcc::String hugeA(4096, 'a');
         EXPECT_STREQ(hugeA.c_str(), msg->GetArg(1)->v_string.str);
     } else {
